feature: 배포 설정
1. 테스트 컨트롤러 작성
2. CorsConfig.java는 기존 config와 겹치는 부분이 있어 삭제
3. Dockerfile base image 변경(리눅스 환경에 맞는 jdk image)
